+++ lib/db_fuel/library.rb
@@ -8,6 +8,7 @@
 #
 
 require_relative 'library/active_record/insert'
+require_relative 'library/active_record/update'
 
 require_relative 'library/dbee/query'
 require_relative 'library/dbee/range'
@@ -16,6 +17,7 @@ module Burner
   # Open up Burner::Jobs and add registrations for this libraries jobs.
   class Jobs
     register 'db_fuel/active_record/insert', DbFuel::Library::ActiveRecord::Insert
+    register 'db_fuel/active_record/update', DbFuel::Library::ActiveRecord::Update
 
     register 'db_fuel/dbee/query', DbFuel::Library::Dbee::Query
     register 'db_fuel/dbee/range', DbFuel::Library::Dbee::Range